check other jobids for unstarted seeds
For each square seed, check if it's already been started, if so continue to the next seed. If not continue as normal and mark the seed as started. When done with its own seeds, check the other job_ids' seeds to see if there are any unstarted, if so then start running those.