New comparison for removing blocks/DMs
For comparing the removal of each individual block from the detector plane I have new code that calculates the sum of the change of the average absolute differences for each energy bin, for each block removed. Likewise we compare the effects of removing of each DM from block 11 compared to the background fit of just block 11.